mds: resolve unsigned coercion compiler warning

+++ src/mds/SessionMap.h
@@ -148,7 +148,7 @@ public:
   interval_set<inodeno_t> pending_prealloc_inos; // journaling prealloc, will be added to prealloc_inos
 
   void notify_cap_release(size_t n_caps);
-  void notify_recall_sent(const int new_limit);
+  void notify_recall_sent(const size_t new_limit);
   void clear_recalled_at();
 
   inodeno_t next_ino() const {
